Adjust StopCall to longer StoryImpl life cycle.
After https://fuchsia-review.googlesource.com/#/c/29707/, StoryImpl
more often than previously can get started after being stopped, which
exposes some bugs in StoryImpl (through user_shell test failing):

* Stop() must Close() the story context binding given to stoey shell,
  so that the binding can be reused in the next Start() call for the
  new story shell instance.

* Stop() may be called during tear down (and in general) on a story
  that is not running, and tolerate that there is no story shell.

* When there is no story shell, the story context binding must not
  be Close()d because it's not bound.

+++ bin/story_runner/story_impl.cc
@@ -243,6 +243,7 @@ class StoryImpl::StopCall : Operation<void> {
   }
 
  private:
+  // StopCall may be run even on a story impl that is not running.
   void Run() {
     // At this point, we don't need to monitor the root modules for state
     // changes anymore, because the next state change of the story is triggered
@@ -280,13 +281,21 @@ class StoryImpl::StopCall : Operation<void> {
   }
 
   void StopStoryShell() {
-    story_impl_->story_shell_->Terminate([this] { StoryShellDown(); });
+    // It StopCall runs on a story that's not running, there is no story shell.
+    if (story_impl_->story_shell_) {
+      story_impl_->story_shell_->Terminate([this] { StoryShellDown(); });
+    } else {
+      StoryShellDown();
+    }
   }
 
   void StoryShellDown() {
     story_impl_->story_shell_controller_.reset();
     story_impl_->story_shell_.reset();
-
+    if (story_impl_->story_context_binding_.is_bound()) {
+      // Close() dchecks if called while not bound.
+      story_impl_->story_context_binding_.Close();
+    }
     StopLinks();
   }
